Decorations2

+++ glass-theme/glass_theme/shaders/lib/fragment_picking.glsl
@@ -10,6 +10,10 @@ vec4 get_picking(int window_id, vec2 scaled_window_coord) {
     // fraction part of float32 is 23 bits. 2^23 = 8388608
     // widget_id is 23*2 - 29 = 17 bits
     // see glass-renderer/view.c:view_pick for the decoding of this
-    return vec4(scaled_window_coord.x, scaled_window_coord.y, (widget_id * 64) + floor(window_id / 8388608), window_id - 8388608 * floor(window_id / 8388608));
+    return vec4(
+      scaled_window_coord.x,
+      scaled_window_coord.y,
+      (widget_id * 64) + int(floor(window_id / 8388608)),
+      window_id - 8388608 * int(floor(window_id / 8388608)));
   }
 }
